# server/src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "World of Stories API",
    "version": "1.0.0",
    "description": "Documentación limpia y modularizada"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
  "paths": {
...
    "/cities/{id}": {
      "get": {
        "operationId": "getCity",
        "tags": [
          "Cities"
        ],
        "summary": "Obtener el detalle profundo de una ciudad",
        "description": "Recupera una ciudad completa por su ID con todos sus arrays hidratados (descriptions, travels, places). Actúa con los virtuals expuestos.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID único de MongoDB de la ciudad",
            "schema": {
              "type": "string",
              "example": "665ed7c4c3f4e567890defab"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle de la ciudad recuperado con éxito.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "city": {
                          "$ref": "#/components/schemas/City"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontró la ciudad especificada."
          }
        }
      },
      "patch": {
        "operationId": "updateCity",
        "tags": [
          "Cities"
        ],
        "summary": "Cambiar estado de publicación de una ciudad",
        "description": "Publica o despublica una ciudad calculando de forma automática la fecha de publicación.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "userName": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado de publicación actualizado correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Ciudad publicada correctamente."
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "city": {
                          "$ref": "#/components/schemas/City"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Ciudad no encontrada."
          }
        }
      },
      "delete": {
        "operationId": "deleteCity",
        "tags": [
          "Cities"
        ],
        "summary": "Eliminar una ciudad por completo",
        "description": "Remueve de forma permanente el documento de la ciudad y todos los subdocumentos anidados (relatos y lugares).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID de la ciudad a destruir",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ciudad eliminada.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "La ciudad \"Luthadel\" y todos sus relatos asociados fueron eliminados correctamente."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontró la ciudad para eliminar."
          }
        }
      }
    },
    "/cities/{id}/publish": {
      "patch": {
        "operationId": "publishCity",
        "tags": [
          "Cities"
        ],
        "summary": "Cambiar estado de publicación de una ciudad",
        "description": "Publica o despublica una ciudad calculando de forma automática la fecha de publicación.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "published"
                ],
                "properties": {
                  "published": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado de publicación actualizado correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Ciudad publicada correctamente."
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "city": {
                          "$ref": "#/components/schemas/City"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Ciudad no encontrada."
          }
        }
      }
    },
    "/cities/{cityId}/description": {
      "patch": {
        "operationId": "addCityDescription",
        "tags": [
          "Cities"
        ],
        "summary": "Añadir una descripción al array de la ciudad",
        "parameters": [
          {
            "name": "cityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "tale",
                  "author"
                ],
                "properties": {
                  "tale": {
                    "type": "Object",
                    "$ref": "#/components/schemas/Readable"
                  },
                  "author": {
                    "type": "string",
                    "example": "Sazed"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Relato añadido.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Descripción añadida correctamente a la ciudad."
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "description": {
                          "$ref": "#/components/schemas/Tale"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateCityDescription",
        "tags": [
          "Cities"
        ],
        "summary": "Actualizar un relato de descripción existente",
        "parameters": [
          {
            "name": "cityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "description"
                ],
                "properties": {
                  "description": {
                    "type": "object",
                    "required": [
                      "id",
                      "tale",
                      "published"
                    ],
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "665ed1b2c3f4e567890abcde"
                      },
                      "tale": {
                        "type": "Object",
                        "$ref": "#/components/schemas/Readable"
                      },
                      "published": {
                        "type": "boolean",
                        "example": true
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Descripción actualizada de forma atómica.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "description": {
                          "$ref": "#/components/schemas/Tale"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/cities/{cityId}/description/{taleId}": {
      "delete": {
        "operationId": "removeCityDescription",
        "tags": [
          "Cities"
        ],
        "summary": "Eliminar una descripción del array",
        "parameters": [
          {
            "name": "cityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Relato removido usando $pull."
          }
        }
      }
    },
...
    "/cities/{cityId}/travel/{taleId}": {
      "delete": {
        "operationId": "removeCityTravel",
        "tags": [
          "Cities"
        ],
        "summary": "Eliminar un viaje del array",
        "parameters": [
          {
            "name": "cityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Relato de viaje eliminado."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Readable": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "example": "El aire soplaba frío sobre las llanuras quebradas..."
          },
          "animation": {
            "type": "string",
            "example": "fade-in-slow"
          }
        }
      },
      "Tale": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "665ed1b2c3f4e567890abcde"
          },
          "tale": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Readable"
            }
          },
          "author": {
            "type": "string",
            "example": "Hoid"
          },
          "published": {
            "type": "boolean",
            "example": false
          },
          "writeDate": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-08T16:24:00.000Z"
          },
          "publishDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": null
          }
        }
      },
      "Place": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "665ed5a1c3f4e567890bcdef"
          },
          "name": {
            "type": "string",
            "example": "Kholinar"
          },
          "published": {
            "type": "boolean",
            "example": true
          },
          "publishDate": {
            "type": "string",
            "format": "date-time",
            "example": "2026-01-15T00:00:00.000Z"
          },
          "events": {
            "type": "array",
            "description": "IDs de los NPCs vinculados a este lugar",
            "items": {
              "type": "string",
              "example": "665ed6b3c3f4e567890cdefa"
            }
          },
          "description": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tale"
            }
          },
          "entry": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tale"
            }
          }
        }
      },
      "City": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "665ed7c4c3f4e567890defab"
          },
          "name": {
            "type": "string",
            "example": "Alethkar"
          },
          "published": {
            "type": "boolean",
            "example": false
          },
          "publishDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "description": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tale"
            }
          },
          "travel": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tale"
            }
          },
          "places": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Place"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}